agent/pyproject.toml
[project]
name = "memory-postcard-agent"
version = "0.1.0"
description = "记忆明信片AI助手 - 集成数据库和消息队列的智能问答系统"
requires-python = ">=3.10"
dependencies = [
    "pocketflow>=0.0.1",
    "openai>=1.0.0",
    "mysql-connector-python>=8.0.0",
    "pika>=1.3.0",
    "python-dotenv>=1.0.0",
    "minio>=7.0.0",
    "fish-audio-sdk>=0.1.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
